build: use warnings when building drumfish
Enabled a slew of warnings that I typically use in my projects (minus
ones that really got mad with simavr headers) and fixed the warnings
that were exposed by them.

+++ src/flash.h
@@ -22,7 +22,7 @@
 #ifndef __FLASH_H__
 #define __FLASH_H__
 
-uint8_t * flash_open_or_create(const char *file, size_t len);
+uint8_t * flash_open_or_create(const char *file, off_t len);
 
 int flash_load(const char *file, uint8_t *start, size_t len);
